+++ DatabaseApi/Services/TherapistServices.cs
@@ -2,6 +2,7 @@
 using DatabaseApi.Models.Entities;
 using Microsoft.Extensions.Options;
 using MongoDB.Driver;
+using ThirdParty.BouncyCastle.Utilities.IO.Pem;
 
 namespace DatabaseApi.Services
 {
@@ -16,6 +17,7 @@ namespace DatabaseApi.Services
 
             _therapists = mongoDatabase.GetCollection<Therapist>(
                 databaseSettings.Value.TherapistCollectionName);
+
         }
 
         public async Task<List<Therapist>> FindAllAsync()
@@ -28,19 +30,29 @@ namespace DatabaseApi.Services
             return await _therapists.Find(x => x.Email == Email).FirstOrDefaultAsync();
         }
 
-        public async Task CreateAsync(Therapist newObject)
+        public async Task<bool> CreateAsync(Therapist newObject)
         {
+            if (await _therapists.Find(c => c.Email == newObject.Email).FirstOrDefaultAsync() != null)
+                return false;
             await _therapists.InsertOneAsync(newObject);
+            return true;
         }
 
-        public async Task UpdateAsync(string Email, Therapist updatedObject)
+        public async Task<bool> UpdateAsync(string Email, Therapist updatedObject)
         {
-            await _therapists.ReplaceOneAsync(x => x.Email == Email, updatedObject);
+            if (await _therapists.Find(c => c.Email == Email).FirstOrDefaultAsync() == null)
+                return false;
+            if (!(await _therapists.ReplaceOneAsync(x => x.Email == Email, updatedObject)).IsAcknowledged)
+                return false;
+            return true;
         }
 
-        public async Task RemoveByIdAsync(string Email)
+        public async Task<bool> RemoveByIdAsync(string Email)
         {
+            if (await _therapists.Find(c => c.Email == Email).FirstOrDefaultAsync() == null)
+                return false;
             await _therapists.DeleteOneAsync(x => x.Email == Email);
+            return true;
         }
 
         public async Task RemoveAllAsync()
@@ -54,19 +66,27 @@ namespace DatabaseApi.Services
             if (therapist == null) return new List<string>();
             return therapist.Patients;
         }
-        public async Task RemovePatient(string Email, string PatientEmail)
+        public async Task<bool> RemovePatient(string Email, string PatientEmail)
         {
-            var therapist = await FindByIdAsync(Email);
-            if (therapist == null) return;
+            var therapist = await _therapists.Find(c => c.Email == Email).FirstOrDefaultAsync();
+            if (therapist == null) return false;
             therapist.Patients.Remove(PatientEmail);
-            await _therapists.ReplaceOneAsync(x => x.Email == Email, therapist);
+            if(!(await _therapists.ReplaceOneAsync(x => x.Email == Email, therapist)).IsAcknowledged) return false;
+            return true;
         }
-        public async Task AssignPatient(string Email, string PatientEmail)
+        public async Task<bool> AssignPatient(string Email, string PatientEmail)
         {
-            var therapist = await FindByIdAsync(Email);
-            if (therapist == null) return;
+            var therapist = await _therapists.Find(c => c.Email == Email).FirstOrDefaultAsync();
+            if (therapist == null) return false;
             therapist.Patients.Add(PatientEmail);
-            await _therapists.ReplaceOneAsync(x => x.Email == Email, therapist);
+            if (!(await _therapists.ReplaceOneAsync(x => x.Email == Email, therapist)).IsAcknowledged) return false;
+            return true;
+        }
+        public async Task<bool> Exists(string Email)
+        {
+            var therapists = await _therapists.Find(c => c.Email == Email).FirstOrDefaultAsync();
+            if (therapists == null) return false;
+            return true;
         }
